+++ atlasdb-client/src/main/java/com/palantir/atlasdb/table/description/TableMetadata.java
@@ -41,10 +41,15 @@ public class TableMetadata implements Persistable {
     final LogSafety nameLogSafety;
 
     public TableMetadata() {
+        this(LogSafety.UNSAFE);
+    }
+
+    public TableMetadata(LogSafety logSafety) {
         this(
                 new NameMetadataDescription(),
                 new ColumnMetadataDescription(),
-                ConflictHandler.RETRY_ON_WRITE_WRITE);
+                ConflictHandler.RETRY_ON_WRITE_WRITE,
+                logSafety);
     }
 
     public TableMetadata(NameMetadataDescription rowMetadata,
@@ -54,12 +59,24 @@ public class TableMetadata implements Persistable {
                 rowMetadata,
                 columns,
                 conflictHandler,
+                LogSafety.UNSAFE);
+    }
+
+    public TableMetadata(NameMetadataDescription rowMetadata,
+            ColumnMetadataDescription columns,
+            ConflictHandler conflictHandler,
+            LogSafety logSafety) {
+        this(
+                rowMetadata,
+                columns,
+                conflictHandler,
                 CachePriority.WARM,
                 false,
                 0,
                 false,
                 SweepStrategy.CONSERVATIVE,
-                false);
+                false,
+                logSafety);
     }
 
     public TableMetadata(NameMetadataDescription rowMetadata,