+++ atlasdb-client/src/main/java/com/palantir/atlasdb/transaction/impl/TransactionConstants.java
@@ -19,6 +19,7 @@ package com.palantir.atlasdb.transaction.impl;
 import com.google.common.collect.ImmutableList;
 import com.palantir.atlasdb.encoding.PtBytes;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
+import com.palantir.atlasdb.protos.generated.TableMetadataPersistence;
 import com.palantir.atlasdb.ptobject.EncodingUtils;
 import com.palantir.atlasdb.table.description.ColumnMetadataDescription;
 import com.palantir.atlasdb.table.description.ColumnValueDescription;
@@ -51,12 +52,14 @@ public class TransactionConstants {
     }
 
     public static final TableMetadata TRANSACTION_TABLE_METADATA = new TableMetadata(
-        NameMetadataDescription.create(ImmutableList.of(new NameComponentDescription.Builder()
-                .componentName("write_ts")
-                .type(ValueType.VAR_LONG)
-                .build())),
-        new ColumnMetadataDescription(ImmutableList.of(
-            new NamedColumnDescription(COMMIT_TS_COLUMN_STRING, "commit_ts", ColumnValueDescription.forType(ValueType.VAR_LONG)))),
-        ConflictHandler.IGNORE_ALL);
+            NameMetadataDescription.create(ImmutableList.of(new NameComponentDescription.Builder()
+                    .componentName("write_ts")
+                    .type(ValueType.VAR_LONG)
+                    .build())),
+            new ColumnMetadataDescription(ImmutableList.of(
+                    new NamedColumnDescription(COMMIT_TS_COLUMN_STRING, "commit_ts",
+                            ColumnValueDescription.forType(ValueType.VAR_LONG)))),
+            ConflictHandler.IGNORE_ALL,
+            TableMetadataPersistence.LogSafety.SAFE);
 
 }