+++ atlasdb-client/src/test/java/com/palantir/atlasdb/table/description/TableMetadataTest.java
@@ -108,12 +108,6 @@ public class TableMetadataTest {
                 NAME_METADATA_DESCRIPTION,
                 COLUMN_METADATA_DESCRIPTION,
                 CONFLICT_HANDLER,
-                TableMetadataPersistence.CachePriority.WARM,
-                false,
-                0,
-                false,
-                TableMetadataPersistence.SweepStrategy.CONSERVATIVE,
-                false,
                 logSafety);
     }
 }